disable the mul_add lint
`mul_add` nukes perf when not built with `-Ctarget_feature=+fma`. maybe I'll
re-add this when `mul_add_relaxed` is stabilized.

+++ generic_daw_gui/src/config.rs
@@ -81,7 +81,7 @@ impl Default for Device {
 #[serde(default)]
 pub struct Autosave {
 	pub enabled: bool,
-	pub interval: NonZero<u64>,
+	pub interval: NonZero<u16>,
 }
 
 impl Default for Autosave {
